drivers/oneplus: goodix: Unify driver code into main driver
