+++ drivers/oneplus/input/fingerprint/goodix/Makefile
@@ -1,1 +1,1 @@
-obj-$(CONFIG_FINGERPRINT_GOODIX)	+= gf_spi.o platform.o netlink.o
+obj-$(CONFIG_FINGERPRINT_GOODIX)	+= gf_spi.o
