+++ drivers/oneplus/input/fingerprint/goodix/gf_spi.c
@@ -45,7 +45,12 @@
 #include <linux/pm_qos.h>
 #include <linux/cpufreq.h>
 //#include <linux/wakelock.h>
+#include <linux/workqueue.h>
 #include <linux/oem/boot_mode.h>
+#include <linux/time.h>
+#include <linux/types.h>
+#include <net/sock.h>
+#include <net/netlink.h>
 #include "gf_spi.h"
 
 #if defined(USE_SPI_BUS)
@@ -56,6 +61,9 @@
 #endif
 
 #include "../fingerprint_detect/fingerprint_detect.h"
+
+#define NETLINK_TEST 25
+#define MAX_MSGSIZE 32
 
 #define VER_MAJOR   1
 #define VER_MINOR   2
@@ -86,6 +94,286 @@ struct gf_key_map maps[] = {
 	{ EV_KEY, GF_KEY_INPUT_BACK },
 	{ EV_KEY, GF_KEY_INPUT_POWER },
 };
+
+static int pid = -1;
+struct sock *gf_nl_sk = NULL;
+
+void sendnlmsg(char *msg)
+{
+	struct sk_buff *skb_1;
+	struct nlmsghdr *nlh;
+	int len = NLMSG_SPACE(MAX_MSGSIZE);
+	int ret = 0;
+	if (!msg || !gf_nl_sk || !pid) {
+		return ;
+	}
+	skb_1 = alloc_skb(len, GFP_KERNEL);
+	if (!skb_1) {
+		pr_err("alloc_skb error\n");
+		return;
+	}
+
+	nlh = nlmsg_put(skb_1, 0, 0, 0, MAX_MSGSIZE, 0);
+
+	NETLINK_CB(skb_1).portid = 0;
+	NETLINK_CB(skb_1).dst_group = 0;
+
+	memcpy(NLMSG_DATA(nlh), msg, sizeof(char));
+	pr_debug("send message: %d\n", *(char *)NLMSG_DATA(nlh));
+
+	ret = netlink_unicast(gf_nl_sk, skb_1, pid, MSG_DONTWAIT);
+	if (!ret) {
+		pr_err("send msg from kernel to usespace failed ret 0x%x\n", ret);
+	}
+}
+
+void sendnlmsg_tp(struct fp_underscreen_info *msg, int length)
+{
+	struct sk_buff *skb_1;
+	struct nlmsghdr *nlh;
+	int len = NLMSG_SPACE(MAX_MSGSIZE);
+	int ret = 0;
+	if (!msg || !gf_nl_sk || !pid) {
+		return ;
+	}
+	skb_1 = alloc_skb(len, GFP_KERNEL);
+	if (!skb_1) {
+		pr_err("alloc_skb error\n");
+		return;
+	}
+
+	nlh = nlmsg_put(skb_1, 0, 0, 0, length, 0);
+
+	NETLINK_CB(skb_1).portid = 0;
+	NETLINK_CB(skb_1).dst_group = 0;
+	memcpy(NLMSG_DATA(nlh), msg, length);//core
+	// pr_debug("send message: %d\n", *(char *)NLMSG_DATA(nlh));
+	ret = netlink_unicast(gf_nl_sk, skb_1, pid, MSG_DONTWAIT);
+	if (!ret) {
+		//kfree_skb(skb_1);
+		pr_err("send msg from kernel to usespace failed ret 0x%x\n", ret);
+	}
+}
+
+void nl_data_ready(struct sk_buff *__skb)
+{
+	struct sk_buff *skb;
+	struct nlmsghdr *nlh;
+	char str[100];
+	skb = skb_get (__skb);
+	if(skb->len >= NLMSG_SPACE(0))
+	{
+		nlh = nlmsg_hdr(skb);
+
+		memcpy(str, NLMSG_DATA(nlh), sizeof(str));
+		pid = nlh->nlmsg_pid;
+
+		kfree_skb(skb);
+	}
+
+}
+
+int netlink_init(void)
+{
+	struct netlink_kernel_cfg netlink_cfg;
+	memset(&netlink_cfg, 0, sizeof(struct netlink_kernel_cfg));
+
+	netlink_cfg.groups = 0;
+	netlink_cfg.flags = 0;
+	netlink_cfg.input = nl_data_ready;
+	netlink_cfg.cb_mutex = NULL;
+
+	gf_nl_sk = netlink_kernel_create(&init_net, NETLINK_TEST,
+			&netlink_cfg);
+
+	if(!gf_nl_sk){
+		pr_err("create netlink socket error\n");
+		return 1;
+	}
+
+	return 0;
+}
+
+void netlink_exit(void)
+{
+	if(gf_nl_sk != NULL){
+		netlink_kernel_release(gf_nl_sk);
+		gf_nl_sk = NULL;
+	}
+
+	pr_info("self module exited\n");
+}
+
+int gf_pinctrl_init(struct gf_dev* gf_dev)
+{
+	int ret = 0;
+	struct device *dev = &gf_dev->spi->dev;
+
+	gf_dev->gf_pinctrl = devm_pinctrl_get(dev);
+	if (IS_ERR_OR_NULL(gf_dev->gf_pinctrl)) {
+		dev_err(dev, "Target does not use pinctrl\n");
+		ret = PTR_ERR(gf_dev->gf_pinctrl);
+		goto err;
+	}
+
+	gf_dev->gpio_state_enable =
+		pinctrl_lookup_state(gf_dev->gf_pinctrl, "fp_en_init");
+	if (IS_ERR_OR_NULL(gf_dev->gpio_state_enable)) {
+		dev_err(dev, "Cannot get active pinstate\n");
+		ret = PTR_ERR(gf_dev->gpio_state_enable);
+		goto err;
+	}
+	if (fp_dtsi_product != 20813) {
+		gf_dev->gpio_state_disable =
+			pinctrl_lookup_state(gf_dev->gf_pinctrl, "fp_dis_init");
+		if (IS_ERR_OR_NULL(gf_dev->gpio_state_disable)) {
+			dev_err(dev, "Cannot get active pinstate\n");
+			ret = PTR_ERR(gf_dev->gpio_state_disable);
+			goto err;
+		}
+}
+	return 0;
+err:
+	gf_dev->gf_pinctrl = NULL;
+	gf_dev->gpio_state_enable = NULL;
+	//gf_dev->gpio_state_disable = NULL;
+	return ret;
+}
+
+int gf_parse_dts(struct gf_dev* gf_dev)
+{
+	int rc = 0;
+	struct device *dev = &gf_dev->spi->dev;
+	struct device_node *np = dev->of_node;
+	//u32 voltage_supply[2];
+	//u32 current_supply;
+
+	gf_dev->reset_gpio = of_get_named_gpio(np, "fp-gpio-reset", 0);
+	if (gf_dev->reset_gpio < 0) {
+		pr_err("falied to get reset gpio!\n");
+		return gf_dev->reset_gpio;
+	}
+
+	rc = devm_gpio_request(dev, gf_dev->reset_gpio, "goodix_reset");
+	if (rc) {
+		pr_err("failed to request reset gpio, rc = %d\n", rc);
+		goto err_reset;
+	}
+	gpio_direction_output(gf_dev->reset_gpio, 0);
+
+	gf_dev->irq_gpio = of_get_named_gpio(np, "fp-gpio-irq", 0);
+	if (gf_dev->irq_gpio < 0) {
+		pr_err("falied to get irq gpio!\n");
+		return gf_dev->irq_gpio;
+	}
+
+	rc = devm_gpio_request(dev, gf_dev->irq_gpio, "goodix_irq");
+	if (rc) {
+		pr_err("failed to request irq gpio, rc = %d\n", rc);
+		goto err_irq;
+	}
+	gpio_direction_input(gf_dev->irq_gpio);
+
+	return rc;
+err_irq:
+	devm_gpio_free(dev, gf_dev->irq_gpio);
+err_reset:
+	devm_gpio_free(dev, gf_dev->reset_gpio);
+	return rc;
+}
+
+void gf_cleanup(struct gf_dev *gf_dev)
+{
+	pr_info("[info] %s\n",__func__);
+	if (gpio_is_valid(gf_dev->irq_gpio))
+	{
+		gpio_free(gf_dev->irq_gpio);
+		pr_info("remove irq_gpio success\n");
+	}
+	if (gpio_is_valid(gf_dev->reset_gpio))
+	{
+		gpio_free(gf_dev->reset_gpio);
+		pr_info("remove reset_gpio success\n");
+	}
+}
+
+int gf_power_on(struct gf_dev* gf_dev)
+{
+	int rc = 0;
+	if (fp_dtsi_product != 20801) {
+		struct device *dev = &gf_dev->spi->dev;
+		struct regulator *vreg = gf_dev->vdd_3v3;
+		if (!vreg) {
+			vreg = regulator_get(dev, "fppower");
+			if (IS_ERR(vreg)) {
+				pr_err("Unable to get fppower power.\n");
+				return PTR_ERR(vreg);
+			}
+		}
+		if (regulator_count_voltages(vreg) > 0) {
+			rc = regulator_set_voltage(vreg, 3324000, 3324000);
+			if (rc) {
+				pr_err("Unable to set voltage on fppower, %d\n", rc);
+			}
+		}
+		rc = regulator_set_load(vreg, 150000);
+		if (rc < 0) {
+			pr_err("Unable to set current on fppower, %d\n", rc);
+		}
+		rc = regulator_enable(vreg);
+		if (rc) {
+			pr_err("error enabling fppower: %d\n", rc);
+			regulator_put(vreg);
+			gf_dev->vdd_3v3  = NULL;
+		}
+	}
+	pr_info("---- power on ok ----\n");
+
+	return rc;
+}
+
+int gf_power_off(struct gf_dev* gf_dev)
+{
+	int rc = 0;
+	if (fp_dtsi_product != 20801) {
+		struct regulator *vreg = gf_dev->vdd_3v3;
+		if (vreg) {
+			if (regulator_is_enabled(vreg)) {
+				regulator_disable(vreg);
+				pr_err("disabled fppower\n");
+			}
+			regulator_put(vreg);
+			gf_dev->vdd_3v3 = NULL;
+		}
+	}
+	pr_info("---- power off ----\n");
+
+	return rc;
+}
+
+int gf_hw_reset(struct gf_dev *gf_dev, unsigned int delay_ms)
+{
+	if(gf_dev == NULL) {
+		pr_info("Input buff is NULL.\n");
+		return -1;
+	}
+	gpio_direction_output(gf_dev->reset_gpio, 1);
+	gpio_set_value(gf_dev->reset_gpio, 0);
+	mdelay(3);
+	gpio_set_value(gf_dev->reset_gpio, 1);
+	mdelay(delay_ms);
+	return 0;
+}
+
+int gf_irq_num(struct gf_dev *gf_dev)
+{
+	if(gf_dev == NULL) {
+		pr_info("Input buff is NULL.\n");
+		return -1;
+	} else {
+		return gpio_to_irq(gf_dev->irq_gpio);
+	}
+}
 
 static void gf_enable_irq(struct gf_dev *gf_dev)
 {
