--- drivers/oneplus/input/fingerprint/goodix/platform.c
@@ -1,194 +1,0 @@
-/*
- * platform indepent driver interface
- *
- * Coypritht (c) 2017 Goodix
- */
-#include <linux/delay.h>
-#include <linux/workqueue.h>
-#include <linux/of_gpio.h>
-#include <linux/gpio.h>
-#include <linux/regulator/consumer.h>
-#include <linux/timer.h>
-#include <linux/err.h>
-
-#include "gf_spi.h"
-#include "../fingerprint_detect/fingerprint_detect.h"
-
-#if defined(USE_SPI_BUS)
-#include <linux/spi/spi.h>
-#include <linux/spi/spidev.h>
-#elif defined(USE_PLATFORM_BUS)
-#include <linux/platform_device.h>
-#endif
-//#include <linux/oneplus/boot_mode.h>
-
-int gf_pinctrl_init(struct gf_dev* gf_dev)
-{
-	int ret = 0;
-	struct device *dev = &gf_dev->spi->dev;
-
-	gf_dev->gf_pinctrl = devm_pinctrl_get(dev);
-	if (IS_ERR_OR_NULL(gf_dev->gf_pinctrl)) {
-		dev_err(dev, "Target does not use pinctrl\n");
-		ret = PTR_ERR(gf_dev->gf_pinctrl);
-		goto err;
-	}
-
-	gf_dev->gpio_state_enable =
-		pinctrl_lookup_state(gf_dev->gf_pinctrl, "fp_en_init");
-	if (IS_ERR_OR_NULL(gf_dev->gpio_state_enable)) {
-		dev_err(dev, "Cannot get active pinstate\n");
-		ret = PTR_ERR(gf_dev->gpio_state_enable);
-		goto err;
-	}
-
-	gf_dev->gpio_state_disable =
-		pinctrl_lookup_state(gf_dev->gf_pinctrl, "fp_dis_init");
-	if (IS_ERR_OR_NULL(gf_dev->gpio_state_disable)) {
-		dev_err(dev, "Cannot get active pinstate\n");
-		ret = PTR_ERR(gf_dev->gpio_state_disable);
-		goto err;
-	}
-
-	return 0;
-err:
-	gf_dev->gf_pinctrl = NULL;
-	gf_dev->gpio_state_enable = NULL;
-	//gf_dev->gpio_state_disable = NULL;
-	return ret;
-}
-int gf_parse_dts(struct gf_dev* gf_dev)
-{
-	int rc = 0;
-	struct device *dev = &gf_dev->spi->dev;
-	struct device_node *np = dev->of_node;
-	//u32 voltage_supply[2];
-	//u32 current_supply;
-
-	gf_dev->reset_gpio = of_get_named_gpio(np, "fp-gpio-reset", 0);
-	if (gf_dev->reset_gpio < 0) {
-		pr_err("falied to get reset gpio!\n");
-		return gf_dev->reset_gpio;
-	}
-
-	rc = devm_gpio_request(dev, gf_dev->reset_gpio, "goodix_reset");
-	if (rc) {
-		pr_err("failed to request reset gpio, rc = %d\n", rc);
-		goto err_reset;
-	}
-	gpio_direction_output(gf_dev->reset_gpio, 0);
-
-	gf_dev->irq_gpio = of_get_named_gpio(np, "fp-gpio-irq", 0);
-	if (gf_dev->irq_gpio < 0) {
-		pr_err("falied to get irq gpio!\n");
-		return gf_dev->irq_gpio;
-	}
-
-	rc = devm_gpio_request(dev, gf_dev->irq_gpio, "goodix_irq");
-	if (rc) {
-		pr_err("failed to request irq gpio, rc = %d\n", rc);
-		goto err_irq;
-	}
-	gpio_direction_input(gf_dev->irq_gpio);
-
-	return rc;
-err_irq:
-	devm_gpio_free(dev, gf_dev->irq_gpio);
-err_reset:
-	devm_gpio_free(dev, gf_dev->reset_gpio);
-	return rc;
-}
-
-void gf_cleanup(struct gf_dev *gf_dev)
-{
-	pr_info("[info] %s\n",__func__);
-	if (gpio_is_valid(gf_dev->irq_gpio))
-	{
-		gpio_free(gf_dev->irq_gpio);
-		pr_info("remove irq_gpio success\n");
-	}
-	if (gpio_is_valid(gf_dev->reset_gpio))
-	{
-		gpio_free(gf_dev->reset_gpio);
-		pr_info("remove reset_gpio success\n");
-	}
-}
-
-int gf_power_on(struct gf_dev* gf_dev)
-{
-	int rc = 0;
-	if (fp_dtsi_product != 20801) {
-		struct device *dev = &gf_dev->spi->dev;
-		struct regulator *vreg = gf_dev->vdd_3v3;
-		if (!vreg) {
-			vreg = regulator_get(dev, "fppower");
-			if (IS_ERR(vreg)) {
-				pr_err("Unable to get fppower power.\n");
-				return PTR_ERR(vreg);
-			}
-		}
-		if (regulator_count_voltages(vreg) > 0) {
-			rc = regulator_set_voltage(vreg, 3024000, 3024000);
-			if (rc) {
-				pr_err("Unable to set voltage on fppower, %d\n", rc);
-			}
-		}
-		rc = regulator_set_load(vreg, 150000);
-		if (rc < 0) {
-			pr_err("Unable to set current on fppower, %d\n", rc);
-		}
-		rc = regulator_enable(vreg);
-		if (rc) {
-			pr_err("error enabling fppower: %d\n", rc);
-			regulator_put(vreg);
-			gf_dev->vdd_3v3  = NULL;
-		}
-	}
-	pr_info("---- power on ok ----\n");
-
-	return rc;
-}
-
-int gf_power_off(struct gf_dev* gf_dev)
-{
-	int rc = 0;
-	if (fp_dtsi_product != 20801) {
-		struct regulator *vreg = gf_dev->vdd_3v3;
-		if (vreg) {
-			if (regulator_is_enabled(vreg)) {
-				regulator_disable(vreg);
-				pr_err("disabled fppower\n");
-			}
-			regulator_put(vreg);
-			gf_dev->vdd_3v3 = NULL;
-		}
-	}
-	pr_info("---- power off ----\n");
-
-	return rc;
-}
-
-int gf_hw_reset(struct gf_dev *gf_dev, unsigned int delay_ms)
-{
-	if(gf_dev == NULL) {
-		pr_info("Input buff is NULL.\n");
-		return -1;
-	}
-	gpio_direction_output(gf_dev->reset_gpio, 1);
-	gpio_set_value(gf_dev->reset_gpio, 0);
-	mdelay(3);
-	gpio_set_value(gf_dev->reset_gpio, 1);
-	mdelay(delay_ms);
-	return 0;
-}
-
-int gf_irq_num(struct gf_dev *gf_dev)
-{
-	if(gf_dev == NULL) {
-		pr_info("Input buff is NULL.\n");
-		return -1;
-	} else {
-		return gpio_to_irq(gf_dev->irq_gpio);
-	}
-}
-
